refactor(common-domain): etc common-core로 파일 위치 변경
